Fix formatting (#61)
* Fix formatting

I had some spare time while waiting for a CEF/Chromium build (...) so I fixed this. As a side effect of the change, I also removed the debug building completely - it just takes time, we don't use the resulting artifacts anyway. So it makes more sense to _only_ build Release builds of this repo.

Closes #60

* Peer suggestions: remove functional changes.