Upgrade to CEF 63
I noted that the CEF 63 binaries are now published. We might as well start doing some of the plumbing for an eventual CefSharp 63.0.0 release.

(The README.txt diff is somewhat more noisy than expected, since I copied in the file from the minimal distribution instead. We are migrating to use that: #57)