Refactor nuspec and targets for swiftshader support
Previously we used a different ItemGroup for every resources/binaries/pak files, we need to add swiftshader support which would have been a 4th target,
all files are always copied in the case of CefSharp, so just switching to one. The ItemGroup was selectively being created on the presence of the
x86 or x64 folder, have created a text file to do the same trick, it's a quick solution, not the cleanest, though can be revisited later without
having to change anything downstream, so going with that for now.